Customize PS1 and and __git_ps1
I am customizing my PS1 as well as adding git information
to it.

First I am making it two lines and rearranging it to largely match
the git bash prompt on windows. The only change is a I added the
time instead of the shell type.

Additionally I am adding a file .git-prompt.sh which will allow
me to add the branch name of the git repo to the commandline
once I enter one.