Add autocmd group
I am adding a autocmd group since this seems like it is best
practice. In it I have a autocmd to allow me to run a Python
script in a terminal which doesn't quite work and a autocmd
for triming trailing white space that I have commented out because
it caused issue for me when I was editing the hunk to try
and make this commit. Irony!