Add filelist and new script to push get remove
I found a bug where the current put script doesn't move the
git-prompt.sh. To avoid these errors in the future I am
adding the files to be controlled in a filelist file. I am
adding a new all in one script that will that will then read
the files from this file list file and get, put, or remove
the files from the home directory.