Fix Windows PS1 prompt bug
When I tried to use my .bashrc on Windows, through the git bash, I
encountered an error where my new line was not being escaped properly.

It turns out that the black slash is interpeted differently in
the git terminal on Windows so I am adjusting that part of the
prompt so that the newline is interpeted correctly.