Set helper for Github mechsin to store
Github will soon require 2FA to be enable by default the result
of this is when pushing by cli you can no longer authenticate.
The solution to this is to use a personal access token, SSH or
PGP keys. I have elected to use the personal access token which
is now stored in a .git-credentials file per machine. The
.git-credentials file is use by the store helper in git so this
change to the .gitconfig tells get to use the store helper when
accessing my personal account on Github.