Add script to install vim packages
I want a way to install all the same Vim packages
on my systems. I came up with this script a while
ago and I am adding it to the repo now. Currently
it only adds one package for indent guides but
I can add more later if I like.