Add a script to setup a user
I am adding and script that will go though and setup
an new user and then copy the configuration files
into the new users home directory. The purpose of this
script is of course to setup my user since these are my
configuration files but it can be used to setup others
as well.