Add different misspelling colors for VIM
The default for VIM misspelled words is very noisy I find the
background color of the text is set to red further there is a
SpellLocal highlighting that uses a cyan color and renders the
words completely illegible in most terminals. So I picked up
some code from stack overflow and added it to my .vimrc so that
now words text is changed to red and underlined red. I would
prefer to have just the red underline but apparently that is a
relatively new feature that isn't in my build and won't work
for all terminals so stick with this for now.